Update all finding the data on NCBI
After going through the lesson, the numbers didn't match up and it would be confusing to show the BioProject Nuber, and then the Run Accessory Table (which is actually the SRA Run Selector Table). This shows the detailed process of going through the data located at BioProject  "PRJNA294072", and gives the instructor an opportunity to "skip" this section if time is short.