replace image with text
Because this image is a text block, it is better to render it as a text block instead of using an image so blind users can understand the content of the block. I've modified the text in two ways:

1. I've replaced the URLs with secure (https) and correct (doi.org) equivalents, but the text will appear the same. 
2. To avoid spamming Dr. Lenski, I modified the email address to remove the `@`. 

This contribution is part of the Core Team accessibility hackathon to add alt-text to images across our lessons. Please feel free to improve upon these suggestions as desired!